+++ build.gradle
@@ -1,0 +1,34 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ *
+ * This generated file contains a sample Java Library project to get you started.
+ * For more details take a look at the Java Libraries chapter in the Gradle
+ * user guide available at https://docs.gradle.org/4.10.2/userguide/java_library_plugin.html
+ */
+
+plugins {
+    id 'java-library'
+}
+
+version = '0.1.7'
+
+dependencies {
+    api 'com.github.javaparser:javaparser-core:3.7.+'
+    testImplementation 'junit:junit:4.12'
+}
+
+repositories {
+    jcenter()
+}
+
+jar {
+    manifest {
+        attributes(
+            'Implementation-Title': project.name,
+            'Implementation-Version': project.version
+        )
+    }
+    from {
+        configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
+    }
+}
