+++ src/main/java/de/thl/jedunit/Constraint.java
@@ -1,0 +1,15 @@
+package de.thl.jedunit;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Restriction annotation.
+ * 
+ * @author Nane Kratzke
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface Constraint { }
