+++ src/test/java/EvaluatorTest.java
@@ -1,0 +1,8 @@
+import org.junit.Test;
+import static org.junit.Assert.*;
+
+public class EvaluatorTest {
+    @Test public void testSomeLibraryMethod() {
+        // TO BE DONE ...
+    }
+}
